vero-agent/pyproject.toml
[project]
name = "vero-agent"
version = "0.1.0"
description = "AI Agent for converting plain English test cases to Vero DSL scripts"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "playwright>=1.49.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
